refactor: extract local role platform mutate to hook (useLocalMutateRolePlatform)
