+++ src/components/[guild]/RoleCard/components/EditRole/hooks/useUpdateRolePlatformVisibility.ts
@@ -4,6 +4,7 @@ import useSubmit from "hooks/useSubmit"
 import useToast from "hooks/useToast"
 import { RolePlatform, Visibility } from "types"
 import { useFetcherWithSign } from "utils/fetcher"
+import useLocalMutateRolePlatform from "./useLocalMutateRolePlatform"
 
 type Props = {
   rolePlatform: RolePlatform
@@ -14,9 +15,10 @@ type Props = {
 type MutateProps = { id: number; visibility: Visibility; visibilityRoleId: number }
 
 const useUpdateRolePlatformVisibility = () => {
-  const { id: guildId, mutateGuild } = useGuild()
+  const { id: guildId } = useGuild()
   const showErrorToast = useShowErrorToast()
   const fetcherWithSign = useFetcherWithSign()
+  const localMutateRolePlatform = useLocalMutateRolePlatform()
   const toast = useToast()
 
   const submit = async (data: Props) => {
@@ -33,57 +35,17 @@ const useUpdateRolePlatformVisibility = () => {
     ])
   }
 
-  const localMutateGuild = (data: MutateProps) => {
-    const { id, visibility, visibilityRoleId } = data
-    mutateGuild(
-      (prevGuild) => ({
-        ...prevGuild,
-        roles: prevGuild.roles.map((role) => {
-          if (role.rolePlatforms.some((rp) => rp.id === id)) {
-            return {
-              ...role,
-              rolePlatforms: findAndUpdateRolePlatform(
-                id,
-                role.rolePlatforms,
-                visibility,
-                visibilityRoleId
-              ),
-            }
-          }
-          return role
-        }),
-      }),
-      { revalidate: false }
-    )
-  }
-
   return useSubmit<Props, MutateProps>(submit, {
     onSuccess: (response) => {
       toast({
         title: "Reward updated!",
         status: "success",
       })
-      localMutateGuild(response)
+      const { id, ...rolePlatformData } = response
+      localMutateRolePlatform(id, rolePlatformData)
     },
     onError: (error) => showErrorToast(error),
   })
 }
-
-const findAndUpdateRolePlatform = (
-  idToUpdate: number,
-  rolePlatforms: RolePlatform[],
-  visibility: Visibility,
-  visibilityRoleId: number
-) =>
-  rolePlatforms.map((rp) => {
-    if (rp.id === idToUpdate) {
-      return {
-        ...rp,
-        visibility,
-        visibilityRoleId,
-      }
-    }
-    return rp
-  })
 
 export default useUpdateRolePlatformVisibility
